@@ -3,28 +3,26 @@ import '../services.dart';
 import 'character_model.dart';
 import 'each_character_route.dart';
 
-List<Character> characters;
+List<Character>? characters;
 
 class CharactersRoute extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    return Scaffold (
+    return Scaffold(
       backgroundColor: Colors.purple[900],
-      body: FutureBuilder<List<Character>> (
+      body: FutureBuilder<List<Character>>(
         future: getAllCharacters(),
         builder: (context, snapshot) {
           if (snapshot.connectionState != ConnectionState.done) {
-            return Center (
-              child: CircularProgressIndicator()
-            );
+            return Center(child: CircularProgressIndicator());
           } else if (snapshot.hasError) {
-            return Center (
-              child: Padding (
+            return Center(
+              child: Padding(
                 padding: EdgeInsets.all(20.0),
                 child: Text(
                   "Please Check Your Internet Connectivity and Try again",
                   textAlign: TextAlign.center,
-                  style: TextStyle (
+                  style: TextStyle(
                     fontSize: 45.0,
                     color: Colors.white,
                   ),
@@ -33,7 +31,7 @@ class CharactersRoute extends StatelessWidget {
             );
           }
           characters = snapshot.data;
-          return Padding (
+          return Padding(
             padding: EdgeInsets.all(10.0),
             child: AllCharacters(),
           );
@@ -48,23 +46,22 @@ class AllCharacters extends StatefulWidget {
 }
 
 class _AllCharacters extends State<AllCharacters> {
-  
   TextEditingController editingController = TextEditingController();
-  var items = List<Character>();
+  var items = <Character>[];
 
   @override
   void initState() {
-    items.addAll(characters);
+    items.addAll(characters!);
     super.initState();
   }
 
   @override
   Widget build(BuildContext context) {
-    return Column (
+    return Column(
       children: <Widget>[
-        Flexible (
+        Flexible(
           flex: 0,
-          child: Padding (
+          child: Padding(
             padding: EdgeInsets.only(top: 15.0, bottom: 10.0),
             child: TextField(
               onChanged: (value) {
@@ -72,53 +69,53 @@ class _AllCharacters extends State<AllCharacters> {
               },
               controller: editingController,
               decoration: InputDecoration(
-                  labelText: "Search",
-                  fillColor: Colors.white,
-                  filled: true,
-                  prefixIcon: Icon(
-                    Icons.search
-                  ),
-                  border: OutlineInputBorder(
-                    borderRadius: BorderRadius.all(Radius.circular(25.0)),
-                  ),
+                labelText: "Search",
+                fillColor: Colors.white,
+                filled: true,
+                prefixIcon: Icon(Icons.search),
+                border: OutlineInputBorder(
+                  borderRadius: BorderRadius.all(Radius.circular(25.0)),
                 ),
+              ),
             ),
           ),
         ),
-        Flexible (
+        Flexible(
           flex: 1,
-          child: ListView.builder (
+          child: ListView.builder(
             itemBuilder: (context, position) {
-              return Padding (
-                padding: EdgeInsets.only(left: 2.0, right: 2.0, top: 1.0, bottom: 1.0),
-                child: GestureDetector (
-                  child: Card (
-                    child: Padding (
+              return Padding(
+                padding: EdgeInsets.only(
+                    left: 2.0, right: 2.0, top: 1.0, bottom: 1.0),
+                child: GestureDetector(
+                  child: Card(
+                    child: Padding(
                       padding: EdgeInsets.all(4.0),
-                      child: Column (
+                      child: Column(
                         children: <Widget>[
                           Text(
-                            '${characters[position].name}',
+                            '${characters![position].name}',
                             textAlign: TextAlign.center,
                             style: TextStyle(
-                                fontSize: 20.0,
-                                fontWeight: FontWeight.bold,
+                              fontSize: 20.0,
+                              fontWeight: FontWeight.bold,
                             ),
                           ),
-                          if (characters[position].role != null) Text(
-                            '${characters[position].role}',
-                            textAlign: TextAlign.center,
-                            style: TextStyle(
+                          if (characters![position].role != null)
+                            Text(
+                              '${characters![position].role}',
+                              textAlign: TextAlign.center,
+                              style: TextStyle(
                                 fontSize: 18.0,
                                 fontStyle: FontStyle.italic,
+                              ),
                             ),
-                          ),
                           Text(
-                            'Blood Status: ${characters[position].bloodStatus}',
+                            'Blood Status: ${characters![position].bloodStatus}',
                             textAlign: TextAlign.center,
                             style: TextStyle(
-                                fontSize: 18.0,
-                                fontStyle: FontStyle.italic,
+                              fontSize: 18.0,
+                              fontStyle: FontStyle.italic,
                             ),
                           ),
                         ],
@@ -126,15 +123,17 @@ class _AllCharacters extends State<AllCharacters> {
                     ),
                   ),
                   onTap: () {
-                    Navigator.push (
+                    Navigator.push(
                       context,
-                      MaterialPageRoute(builder: (context) => EachCharacterRoute(characterName: characters[position].name)),
+                      MaterialPageRoute(
+                          builder: (context) => EachCharacterRoute(
+                              characterName: characters![position].name)),
                     );
                   },
                 ),
               );
             },
-            itemCount: characters.length,
+            itemCount: characters!.length,
           ),
         ),
       ],
@@ -142,12 +141,12 @@ class _AllCharacters extends State<AllCharacters> {
   }
 
   void filterSearchResults(String query) {
-    List<Character> dummySearchList = List<Character>();
-    dummySearchList.addAll(characters);
-    if(query.isNotEmpty) {
-      List<Character> dummyListData = List<Character>();
+    List<Character> dummySearchList = <Character>[];
+    dummySearchList.addAll(characters!);
+    if (query.isNotEmpty) {
+      List<Character> dummyListData = <Character>[];
       dummySearchList.forEach((item) {
-        if(item.name.contains(query)) {
+        if (item.name!.contains(query)) {
           dummyListData.add(item);
         }
       });
@@ -159,9 +158,8 @@ class _AllCharacters extends State<AllCharacters> {
     } else {
       setState(() {
         items.clear();
-        items.addAll(characters);
+        items.addAll(characters!);
       });
     }
   }
-
 }
