@@ -4,11 +4,11 @@ import 'house_model.dart';
 import '../characters/each_character_route.dart';
 
 class EachHouseRoute extends StatelessWidget {
-  
   final String houseName;
-  final Color backColor;
+  final Color? backColor;
 
-  EachHouseRoute({Key key, @required this.houseName, @required this.backColor}) : super(key: key);
+  EachHouseRoute({Key? key, required this.houseName, required this.backColor})
+      : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -18,17 +18,15 @@ class EachHouseRoute extends StatelessWidget {
         future: getParticularHouse(houseName),
         builder: (context, snapshot) {
           if (snapshot.connectionState != ConnectionState.done) {
-            return Center (
-              child: CircularProgressIndicator()
-            );
+            return Center(child: CircularProgressIndicator());
           } else if (snapshot.hasError) {
-            return Center (
-              child: Padding (
+            return Center(
+              child: Padding(
                 padding: EdgeInsets.all(20.0),
                 child: Text(
                   "Please Check Your Internet Connectivity and Try again",
                   textAlign: TextAlign.center,
-                  style: TextStyle (
+                  style: TextStyle(
                     fontSize: 45.0,
                     color: Colors.white,
                   ),
@@ -36,103 +34,122 @@ class EachHouseRoute extends StatelessWidget {
               ),
             );
           }
-          return Padding (
+          return Padding(
             padding: EdgeInsets.all(10.0),
-            child: Column (
+            child: Column(
               children: <Widget>[
-                Flexible (
+                Flexible(
                   flex: 0,
-                  child: Card (
-                    child: Padding (
+                  child: Card(
+                    child: Padding(
                       padding: EdgeInsets.all(15.0),
-                      child: Column (
+                      child: Column(
                         mainAxisSize: MainAxisSize.min,
                         crossAxisAlignment: CrossAxisAlignment.stretch,
                         children: <Widget>[
                           Text(
-                            '${snapshot.data[0].name}',
+                            '${snapshot.data![0].name}',
                             style: TextStyle(
-                                fontSize: 40.0,
-                                fontWeight: FontWeight.bold,
+                              fontSize: 40.0,
+                              fontWeight: FontWeight.bold,
                             ),
                           ),
-                          if (snapshot.data[0].mascot != null) RichText(
-                            text: new TextSpan (
-                              children: <TextSpan>[
-                                TextSpan (text: 'Mascot: ', style: TextStyle(fontWeight: FontWeight.bold)),
-                                TextSpan (text: '${snapshot.data[0].mascot}'),
-                              ],
-                              style: TextStyle(
+                          if (snapshot.data![0].mascot != null)
+                            RichText(
+                              text: new TextSpan(
+                                children: <TextSpan>[
+                                  TextSpan(
+                                      text: 'Mascot: ',
+                                      style: TextStyle(
+                                          fontWeight: FontWeight.bold)),
+                                  TextSpan(text: '${snapshot.data![0].mascot}'),
+                                ],
+                                style: TextStyle(
                                   fontSize: 20.0,
                                   color: Colors.black,
                                   fontStyle: FontStyle.italic,
+                                ),
                               ),
                             ),
-                          ),
-                          if (snapshot.data[0].headOfHouse != null) RichText(
-                            text: new TextSpan (
-                              children: <TextSpan>[
-                                TextSpan (text: 'Head of House: ', style: TextStyle(fontWeight: FontWeight.bold)),
-                                TextSpan (text: '${snapshot.data[0].headOfHouse}'),
-                              ],
-                              style: TextStyle(
+                          if (snapshot.data![0].headOfHouse != null)
+                            RichText(
+                              text: new TextSpan(
+                                children: <TextSpan>[
+                                  TextSpan(
+                                      text: 'Head of House: ',
+                                      style: TextStyle(
+                                          fontWeight: FontWeight.bold)),
+                                  TextSpan(
+                                      text: '${snapshot.data![0].headOfHouse}'),
+                                ],
+                                style: TextStyle(
                                   fontSize: 20.0,
                                   color: Colors.black,
                                   fontStyle: FontStyle.italic,
+                                ),
                               ),
                             ),
-                          ),
-                          if (snapshot.data[0].houseGhost != null) RichText(
-                            text: new TextSpan (
-                              children: <TextSpan>[
-                                TextSpan (text: 'House Ghost: ', style: TextStyle(fontWeight: FontWeight.bold)),
-                                TextSpan (text: '${snapshot.data[0].houseGhost}'),
-                              ],
-                              style: TextStyle(
+                          if (snapshot.data![0].houseGhost != null)
+                            RichText(
+                              text: new TextSpan(
+                                children: <TextSpan>[
+                                  TextSpan(
+                                      text: 'House Ghost: ',
+                                      style: TextStyle(
+                                          fontWeight: FontWeight.bold)),
+                                  TextSpan(
+                                      text: '${snapshot.data![0].houseGhost}'),
+                                ],
+                                style: TextStyle(
                                   fontSize: 20.0,
                                   color: Colors.black,
                                   fontStyle: FontStyle.italic,
+                                ),
                               ),
                             ),
-                          ),
-                          if (snapshot.data[0].founder != null) RichText(
-                            text: new TextSpan (
-                              children: <TextSpan>[
-                                TextSpan (text: 'Founder: ', style: TextStyle(fontWeight: FontWeight.bold)),
-                                TextSpan (text: '${snapshot.data[0].founder}'),
-                              ],
-                              style: TextStyle(
+                          if (snapshot.data![0].founder != null)
+                            RichText(
+                              text: new TextSpan(
+                                children: <TextSpan>[
+                                  TextSpan(
+                                      text: 'Founder: ',
+                                      style: TextStyle(
+                                          fontWeight: FontWeight.bold)),
+                                  TextSpan(
+                                      text: '${snapshot.data![0].founder}'),
+                                ],
+                                style: TextStyle(
                                   fontSize: 20.0,
                                   color: Colors.black,
                                   fontStyle: FontStyle.italic,
+                                ),
                               ),
                             ),
-                          ),
                         ],
                       ),
                     ),
                   ),
                 ),
-                Flexible (
+                Flexible(
                   flex: 1,
-                  child: ListView.builder (
+                  child: ListView.builder(
                     scrollDirection: Axis.vertical,
                     shrinkWrap: true,
                     itemBuilder: (context, position) {
-                      return Padding (
+                      return Padding(
                         padding: EdgeInsets.only(top: 1.0, bottom: 1.0),
-                        child: GestureDetector (
-                          child: Card (
-                            child: Padding (
+                        child: GestureDetector(
+                          child: Card(
+                            child: Padding(
                               padding: EdgeInsets.all(7.0),
-                              child: Column (
+                              child: Column(
                                 children: <Widget>[
                                   Text(
-                                    '${snapshot.data[0].members[position].name}',
+                                    '${snapshot.data![0].members![position].name}',
                                     textAlign: TextAlign.center,
                                     style: TextStyle(
-                                        fontSize: 23.0,
-                                        fontWeight: FontWeight.bold,
+                                      fontSize: 23.0,
+                                      fontWeight: FontWeight.bold,
                                     ),
                                   ),
                                 ],
@@ -140,29 +157,41 @@ class EachHouseRoute extends StatelessWidget {
                             ),
                           ),
                           onTap: () {
-                            Navigator.push (
+                            Navigator.push(
                               context,
-                              MaterialPageRoute(builder: (context) => EachCharacterRoute(characterName: snapshot.data[0].members[position].name)),
+                              MaterialPageRoute(
+                                  builder: (context) => EachCharacterRoute(
+                                      characterName: snapshot
+                                          .data![0].members![position].name)),
                             );
                           },
                         ),
                       );
                     },
-                    itemCount: snapshot.data[0].members.length,
+                    itemCount: snapshot.data![0].members!.length,
                   ),
                 ),
               ],
             ),
           );
         },
-      ), 
+      ),
     );
   }
 
   Future<List<House>> getParticularHouse(String name) {
-    if(name == "Gryffindor") {return getGryffindor();}
-    if(name == "Slytherin") {return getSlytherin();}
-    if(name == "Hufflepuff") {return getHufflepuff();}
-    if(name == "Ravenclaw") {return getRavenclaw();}
+    if (name == "Gryffindor") {
+      return getGryffindor();
+    }
+    if (name == "Slytherin") {
+      return getSlytherin();
+    }
+    if (name == "Hufflepuff") {
+      return getHufflepuff();
+    }
+    if (name == "Ravenclaw") {
+      return getRavenclaw();
+    }
+    return getGryffindor();
   }
 }
