@@ -4,72 +4,88 @@ import 'each_house_route.dart';
 class HousesRoute extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    return Container (
-      decoration: BoxDecoration (
-        image: DecorationImage (
+    return Container(
+      decoration: BoxDecoration(
+        image: DecorationImage(
           image: AssetImage('assets/images/houses.jpg'),
           fit: BoxFit.cover,
         ),
       ),
-      child: Scaffold (
+      child: Scaffold(
         backgroundColor: Colors.transparent,
-        body: Column (
+        body: Column(
           children: <Widget>[
-            Flexible (
+            Flexible(
               flex: 1,
-              child: Row (
+              child: Row(
                 children: <Widget>[
-                  Flexible (
+                  Flexible(
                     flex: 1,
-                    child: Opacity (
+                    child: Opacity(
                       opacity: 0.7,
-                      child: Padding (
+                      child: Padding(
                         padding: EdgeInsets.all(4.0),
-                        child: SizedBox.expand (
-                          child: RaisedButton (
-                            color: Colors.red[900],
-                            shape: new RoundedRectangleBorder(borderRadius: new BorderRadius.circular(20.0)),
-                            child: Text(
-                              "Gryffindor",
-                              style: TextStyle(
-                                color: Colors.white,
-                                fontSize: 25.0,
-                              ),
+                        child: SizedBox.expand(
+                          child: Container(
+                            decoration: BoxDecoration(
+                              borderRadius: new BorderRadius.circular(20.0),
                             ),
-                            onPressed: () {
-                              Navigator.push (
-                                context,
-                                MaterialPageRoute(builder: (context) => EachHouseRoute(houseName: "Gryffindor", backColor: Colors.red[900])),
-                              );
-                            },
+                            child: ElevatedButton(
+                              // color: Colors.red[900],
+                              child: Text(
+                                "Gryffindor",
+                                style: TextStyle(
+                                  color: Colors.white,
+                                  fontSize: 25.0,
+                                ),
+                              ),
+                              onPressed: () {
+                                Navigator.push(
+                                  context,
+                                  MaterialPageRoute(
+                                      builder: (context) => EachHouseRoute(
+                                          houseName: "Gryffindor",
+                                          backColor: Colors.red[900])),
+                                );
+                              },
+                            ),
                           ),
                         ),
                       ),
                     ),
                   ),
-                  Flexible (
+                  Flexible(
                     flex: 1,
-                    child: Opacity (
+                    child: Opacity(
                       opacity: 0.7,
-                      child: Padding (
+                      child: Padding(
                         padding: EdgeInsets.all(4.0),
-                        child: SizedBox.expand (
-                          child: RaisedButton (
-                            color: Colors.green[900],
-                            shape: new RoundedRectangleBorder(borderRadius: new BorderRadius.circular(20.0)),
-                            child: Text(
-                              "Slytherin",
-                              style: TextStyle(
-                                color: Colors.white,
-                                fontSize: 25.0,
-                              ),
+                        child: SizedBox.expand(
+                          child: Container(
+                            decoration: BoxDecoration(
+                              borderRadius: new BorderRadius.circular(20.0),
                             ),
-                            onPressed: () {
-                              Navigator.push (
-                                context,
-                                MaterialPageRoute(builder: (context) => EachHouseRoute(houseName: "Slytherin", backColor: Colors.green[900])),
-                              );
-                            },
+                            child: ElevatedButton(
+                              // color: Colors.green[900],
+                              // shape: new RoundedRectangleBorder(
+                              //     borderRadius: new BorderRadius.circular(20.0)),
+                              child: Text(
+                                "Slytherin",
+                                style: TextStyle(
+                                  color: Colors.white,
+                                  fontSize: 25.0,
+                                ),
+                              ),
+                              onPressed: () {
+                                Navigator.push(
+                                  context,
+                                  MaterialPageRoute(
+                                      builder: (context) => EachHouseRoute(
+                                          houseName: "Slytherin",
+                                          backColor: Colors.green[900])),
+                                );
+                              },
+                            ),
                           ),
                         ),
                       ),
@@ -78,61 +94,79 @@ class HousesRoute extends StatelessWidget {
                 ],
               ),
             ),
-            Flexible (
+            Flexible(
               flex: 1,
-              child: Row (
+              child: Row(
                 children: <Widget>[
-                  Flexible (
+                  Flexible(
                     flex: 1,
-                    child: Opacity (
+                    child: Opacity(
                       opacity: 0.7,
-                      child: Padding (
+                      child: Padding(
                         padding: EdgeInsets.all(4.0),
-                        child: SizedBox.expand (
-                          child: RaisedButton (
-                            color: Colors.yellow[700],
-                            shape: new RoundedRectangleBorder(borderRadius: new BorderRadius.circular(20.0)),
-                            child: Text(
-                              "Hufflepuff",
-                              style: TextStyle(
-                                color: Colors.white,
-                                fontSize: 25.0,
-                              ),
+                        child: SizedBox.expand(
+                          child: Container(
+                            decoration: BoxDecoration(
+                              borderRadius: new BorderRadius.circular(20.0),
                             ),
-                            onPressed: () {
-                              Navigator.push (
-                                context,
-                                MaterialPageRoute(builder: (context) => EachHouseRoute(houseName: "Hufflepuff", backColor: Colors.yellow[700])),
-                              );
-                            },
+                            child: ElevatedButton(
+                              // color: Colors.yellow[700],
+                              // shape: new RoundedRectangleBorder(
+                              //     borderRadius: new BorderRadius.circular(20.0)),
+                              child: Text(
+                                "Hufflepuff",
+                                style: TextStyle(
+                                  color: Colors.white,
+                                  fontSize: 25.0,
+                                ),
+                              ),
+                              onPressed: () {
+                                Navigator.push(
+                                  context,
+                                  MaterialPageRoute(
+                                      builder: (context) => EachHouseRoute(
+                                          houseName: "Hufflepuff",
+                                          backColor: Colors.yellow[700])),
+                                );
+                              },
+                            ),
                           ),
                         ),
                       ),
                     ),
                   ),
-                  Flexible (
+                  Flexible(
                     flex: 1,
-                    child: Opacity (
+                    child: Opacity(
                       opacity: 0.7,
-                      child: Padding (
+                      child: Padding(
                         padding: EdgeInsets.all(4.0),
-                        child: SizedBox.expand (
-                          child: RaisedButton (
-                            color: Colors.blue[900],
-                            shape: new RoundedRectangleBorder(borderRadius: new BorderRadius.circular(20.0)),
-                            child: Text(
-                              "Ravenclaw",
-                              style: TextStyle(
-                                color: Colors.white,
-                                fontSize: 25.0,
-                              ),
+                        child: SizedBox.expand(
+                          child: Container(
+                            decoration: BoxDecoration(
+                              borderRadius: new BorderRadius.circular(20.0),
                             ),
-                            onPressed: () {
-                              Navigator.push (
-                                context,
-                                MaterialPageRoute(builder: (context) => EachHouseRoute(houseName: "Ravenclaw", backColor: Colors.blue[900])),
-                              );
-                            },
+                            child: ElevatedButton(
+                              // color: Colors.blue[900],
+                              // shape: new RoundedRectangleBorder(
+                              //     borderRadius: new BorderRadius.circular(20.0)),
+                              child: Text(
+                                "Ravenclaw",
+                                style: TextStyle(
+                                  color: Colors.white,
+                                  fontSize: 25.0,
+                                ),
+                              ),
+                              onPressed: () {
+                                Navigator.push(
+                                  context,
+                                  MaterialPageRoute(
+                                      builder: (context) => EachHouseRoute(
+                                          houseName: "Ravenclaw",
+                                          backColor: Colors.blue[900])),
+                                );
+                              },
+                            ),
                           ),
                         ),
                       ),
