@@ -6,16 +6,16 @@ import 'houses/houses_route.dart';
 import 'spells/spells_route.dart';
 
 void main() {
-  SystemChrome.setPreferredOrientations([DeviceOrientation.portraitUp]).then((_) {
-    runApp(HarryPotterApp());
-  });
+  // SystemChrome.setPreferredOrientations([DeviceOrientation.portraitUp])
+  //     .then((_) {
+  runApp(HarryPotterApp());
+  // });
 }
 
 class HarryPotterApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
-        home: HomeScreen());
+    return MaterialApp(home: HomeScreen());
   }
 }
 
@@ -30,9 +30,9 @@ class HomeScreen extends StatelessWidget {
         accentColor: Colors.amber,
         fontFamily: 'Montserrat',
       ),
-      home: Container (
-        decoration: BoxDecoration (
-          image: DecorationImage (
+      home: Container(
+        decoration: BoxDecoration(
+          image: DecorationImage(
             image: AssetImage('assets/images/hogwarts.jpg'),
             fit: BoxFit.cover,
           ),
@@ -43,8 +43,8 @@ class HomeScreen extends StatelessWidget {
             mainAxisAlignment: MainAxisAlignment.spaceEvenly,
             crossAxisAlignment: CrossAxisAlignment.center,
             children: <Widget>[
-              Center (
-                child: Text (
+              Center(
+                child: Text(
                   'Harry Potter App',
                   style: TextStyle(
                     fontSize: 35.0,
@@ -52,106 +52,132 @@ class HomeScreen extends StatelessWidget {
                   ),
                 ),
               ),
-              Center (
-                child: Opacity (
+              Center(
+                child: Opacity(
                   opacity: 0.7,
-                  child: SizedBox (
+                  child: SizedBox(
                     width: 200.0,
                     height: 80.0,
-                    child: RaisedButton (
-                      shape: new RoundedRectangleBorder(borderRadius: new BorderRadius.circular(20.0)),
-                      color: Colors.blue[900],
-                      child: Text(
-                        "Hat",
-                        style: TextStyle(
-                          color: Colors.white,
-                          fontSize: 30.0,
-                        ),
+                    child: Container(
+                      decoration: BoxDecoration(
+                        borderRadius: new BorderRadius.circular(20.0),
                       ),
-                      onPressed: () {
-                        Navigator.push (
-                          context,
-                          MaterialPageRoute(builder: (context) => HatRoute()),
-                        );
-                      },
+                      child: ElevatedButton(
+                        // shape: new RoundedRectangleBorder(borderRadius: new BorderRadius.circular(20.0)),
+                        // color: Colors.blue[900],
+                        child: Text(
+                          "Hat",
+                          style: TextStyle(
+                            color: Colors.white,
+                            fontSize: 30.0,
+                          ),
+                        ),
+                        onPressed: () {
+                          Navigator.push(
+                            context,
+                            MaterialPageRoute(builder: (context) => HatRoute()),
+                          );
+                        },
+                      ),
                     ),
                   ),
                 ),
               ),
-              Center (
-                child: Opacity (
+              Center(
+                child: Opacity(
                   opacity: 0.7,
-                  child: SizedBox (
+                  child: SizedBox(
                     width: 200.0,
                     height: 80.0,
-                    child: RaisedButton (
-                      shape: new RoundedRectangleBorder(borderRadius: new BorderRadius.circular(20.0)),
-                      color: Colors.blue[900],
-                      child: Text(
-                        "Characters",
-                        style: TextStyle(
-                          color: Colors.white,
-                          fontSize: 30.0,
-                        ),
+                    child: Container(
+                      decoration: BoxDecoration(
+                        borderRadius: new BorderRadius.circular(20.0),
                       ),
-                      onPressed: () {
-                        Navigator.push (
-                          context,
-                          MaterialPageRoute(builder: (context) => CharactersRoute()),
-                        );
-                      },
+                      child: ElevatedButton(
+                        // shape: new RoundedRectangleBorder(
+                        //     borderRadius: new BorderRadius.circular(20.0)),
+                        // color: Colors.blue[900],
+                        child: Text(
+                          "Characters",
+                          style: TextStyle(
+                            color: Colors.white,
+                            fontSize: 30.0,
+                          ),
+                        ),
+                        onPressed: () {
+                          Navigator.push(
+                            context,
+                            MaterialPageRoute(
+                                builder: (context) => CharactersRoute()),
+                          );
+                        },
+                      ),
                     ),
                   ),
                 ),
               ),
-              Center (
-                child: Opacity (
+              Center(
+                child: Opacity(
                   opacity: 0.7,
-                  child: SizedBox (
+                  child: SizedBox(
                     width: 200.0,
                     height: 80.0,
-                    child: RaisedButton (
-                      shape: new RoundedRectangleBorder(borderRadius: new BorderRadius.circular(20.0)),
-                      color: Colors.blue[900],
-                      child: Text(
-                        "Houses",
-                        style: TextStyle(
-                          color: Colors.white,
-                          fontSize: 30.0,
-                        ),
+                    child: Container(
+                      decoration: BoxDecoration(
+                        borderRadius: new BorderRadius.circular(20.0),
                       ),
-                      onPressed: () {
-                        Navigator.push (
-                          context,
-                          MaterialPageRoute(builder: (context) => HousesRoute()),
-                        );
-                      },
+                      child: ElevatedButton(
+                        // shape: new RoundedRectangleBorder(
+                        //     borderRadius: new BorderRadius.circular(20.0)),
+                        // color: Colors.blue[900],
+                        child: Text(
+                          "Houses",
+                          style: TextStyle(
+                            color: Colors.white,
+                            fontSize: 30.0,
+                          ),
+                        ),
+                        onPressed: () {
+                          Navigator.push(
+                            context,
+                            MaterialPageRoute(
+                                builder: (context) => HousesRoute()),
+                          );
+                        },
+                      ),
                     ),
                   ),
                 ),
               ),
-              Center (
-                child: Opacity (
+              Center(
+                child: Opacity(
                   opacity: 0.7,
-                  child: SizedBox (
+                  child: SizedBox(
                     width: 200.0,
                     height: 80.0,
-                    child: RaisedButton (
-                      shape: new RoundedRectangleBorder(borderRadius: new BorderRadius.circular(20.0)),
-                      color: Colors.blue[900],
-                      child: Text(
-                        "Spells",
-                        style: TextStyle(
-                          color: Colors.white,
-                          fontSize: 30.0,
-                        ),
+                    child: Container(
+                      decoration: BoxDecoration(
+                        borderRadius: new BorderRadius.circular(20.0),
                       ),
-                      onPressed: () {
-                        Navigator.push (
-                          context,
-                          MaterialPageRoute(builder: (context) => SpellsRoute()),
-                        );
-                      },
+                      child: ElevatedButton(
+                        // shape: new RoundedRectangleBorder(
+                        //     borderRadius: new BorderRadius.circular(20.0)),
+                        // color: Colors.blue[900],
+                        child: Text(
+                          "Spells",
+                          style: TextStyle(
+                            color: Colors.white,
+                            fontSize: 30.0,
+                          ),
+                        ),
+                        onPressed: () {
+                          Navigator.push(
+                            context,
+                            MaterialPageRoute(
+                                builder: (context) => SpellsRoute()),
+                          );
+                        },
+                      ),
                     ),
                   ),
                 ),
